feat(discovery): add HT + group shortcuts to the standalone Sub page
A loose Sub is most useful bonded into a home theater or a speaker group, so the
sub page now offers the same "put it to use" shortcuts the room page has. Extract
the shared shortcut pieces (ActionRow, leaveTo, addToHomeTheater, homeTheaterTargets)
into bonding_shortcuts.dart so RoomScreen and SubScreen share one implementation.
